making pages more public

+++ app/models/favicon.rb
@@ -1,4 +1,6 @@
 class Favicon < ActiveRecord::Base
+  attr_accessor :skip_load
+
   has_attached_file :favicon,
                     :styles => { 
                       :small => { :geometry => "16x16>", :format => 'png' },
@@ -7,11 +9,11 @@ class Favicon < ActiveRecord::Base
                     :storage => :s3,
                     :s3_credentials => "#{Rails.root}/config/s3.yml",
                     :path => "/favicons/:uid.:style.:extension"
-  validates_presence_of :original_url
+  validates_presence_of :url
 
   composed_of :uri,
-    :class_name => 'FriendlyUrl',
-    :mapping    => %w(original_url to_s),
+    :class_name => 'HandyUrl',
+    :mapping    => %w(url to_s),
     :allow_nil  => true,
     :converter  => :new
 
@@ -22,31 +24,30 @@ class Favicon < ActiveRecord::Base
       favicon = "/#{favicon}" unless favicon[0] == "/"
       parsed_favicon_url = "#{uri.scheme}://#{uri.host}#{favicon}"
     end
-    favicon = Favicon.find_by_original_url(parsed_favicon_url)
-    favicon || Favicon.create({:original_url => parsed_favicon_url, :uid => uri.uid})
+    favicon = Favicon.find_by_url(parsed_favicon_url)
+    favicon || Favicon.create({ :url => parsed_favicon_url, :uid => uri.uid })
   end
 
   def download_favicon!
     Timeout::timeout(20) do
       fav = self.favicon = download_remote_image
-      if fav.nil? || !save
-        Favicon.delete_all "id=#{self.id}"
-      end
+      # If not able to save the favicon, just delete it from favicons
+      Favicon.delete_all "id=#{self.id}" if fav.nil? || !save
     end
   end
   handle_asynchronously :download_favicon!
-  after_save :download_favicon!, :if => :original_url_changed?
+  after_save :download_favicon!, :if => :should_download?
   
-  private  
-  def original_url_provided?
-    !self.original_url.blank?
-  end
-
+  private
   def download_remote_image
-    io = open(URI.parse(original_url))
+    io = open(URI.parse(url))
     def io.original_filename; base_uri.path.split('/').last; end
     io.original_filename.blank? ? nil : io
   rescue # catch url errors with validations instead of exceptions (Errno::ENOENT, OpenURI::HTTPError, etc...)
+  end
+  
+  def should_download?
+    url_changed? && !skip_load
   end
   
   after_save :update_associated_items, :if => :favicon_file_name_changed?
